requirements.txt
amqp==2.3.2
Babel==2.6.0
billiard==3.5.0.3
celery==4.2.0
django-celery-results==1.0.1
flower==0.9.2
kombu==4.2.1
pytz==2018.4
SQLAlchemy==1.2.8
tornado==5.0.2
vine==1.1.4
